+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - Simplified layout of GIF display
 - Simplify log output
 
+### Fixed
+- Proper catching of imgur client errors
+
 ## [0.0.0] - 2015-09-02
 Initial release